add middleware (#10)
* add initial data and delete all data

* add authentication

* add middleware

+++ api/controllers/attendance.js
@@ -1,6 +1,6 @@
 import Attendance from "../models/attendance.js";
 import mongoose from "mongoose";
-import { joiClockInSchema } from "../utils/joivalidate.js"
+import Employee from "../models/employee.js";
 /**
  * @openapi
  * paths:
@@ -10,6 +10,8 @@ import { joiClockInSchema } from "../utils/joivalidate.js"
  *       description: Retrieve **attendances** by employee username.
  *       tags:
  *         - Attendance
+ *       security:
+ *        - jwt: []
  *       parameters:
  *         - name: username
  *           in: path
@@ -35,6 +37,27 @@ import { joiClockInSchema } from "../utils/joivalidate.js"
  *                 $ref: '#/components/schemas/ErrorMessage'
  *               example:
  *                 message: "Query parameter 'username' is required"
+ *         '401':
+ *            description: <b>Unauthorized</b>, with error message.
+ *            content:
+ *              application/json:
+ *               schema:
+ *                 $ref: '#/components/schemas/ErrorMessage'
+ *               examples:
+ *                no token provided:
+ *                 value:
+ *                  message: No authorization token was found.
+ *                user not found:
+ *                 value:
+ *                  message: User not found.
+ *         '403':
+ *          description: <b>Forbidden</b>, with error message.
+ *          content:
+ *           application/json:
+ *            schema:
+ *             $ref: '#/components/schemas/ErrorMessage'
+ *            example:
+ *             message: Not authorized to access this info.
  *         '500':
  *           description: <b>Internal Server Error</b>, with error message.
  *           content:
@@ -46,46 +69,45 @@ import { joiClockInSchema } from "../utils/joivalidate.js"
  */
 
 const attendanceByUsername = async (req, res) => {
-  try {
-    if (
-      !req.params.username ||
-      req.params.username == undefined ||
-      !req.params.username.length < 0
-    ) {
-      res.status(400).json({
-        message: "Query parameter 'username' is required",
-      });
-      return;
-    }
-    const attendances = await Attendance.find({
-      userName: req.params.username,
-    }).exec();
+  getEmployee(req, res, async (req, res, emp) => {
+    try {
+      if (
+        !req.params.username ||
+        req.params.username == undefined ||
+        !req.params.username.length < 0
+      ) {
+        res.status(400).json({
+          message: "Query parameter 'username' is required",
+        });
+        return;
+      }
+      if(emp.role != "admin" && emp.userName != req.params.username){
+        return res.status(403).json({
+          message: "Not authorized to access this info.",
+        });
+      }
+      const attendances = await Attendance.find({
+        userName: req.params.username,
+      }).exec();
 
-    res.status(200).json(attendances);
-  } catch (err) {
-    res.status(500).json({ message: err.message });
-  }
+      res.status(200).json(attendances);
+    } catch (err) {
+      res.status(500).json({ message: err.message });
+    }
+  });
 };
 
 /**
  * @openapi
  * paths:
- *   /clockIn/{username}:
+ *   /clockIn:
  *     post:
  *       summary: Clockin employee.
  *       description: Insert attendace with current date-time.
  *       tags:
  *         - Attendance
- *       parameters:
- *         - name: username
- *           in: path
- *           required: true
- *           description: <b>username</b> of the employee
- *           schema:
- *             type: string
- *             minLength: 0
- *             maxLength: 100
- *           example: admin
+ *       security:
+ *        - jwt: []
  *       responses:
  *         '201':
  *           description: <b>OK</b>, with created attendance.
@@ -101,6 +123,27 @@ const attendanceByUsername = async (req, res) => {
  *                 $ref: '#/components/schemas/ErrorMessage'
  *               example:
  *                 message: "Query parameter 'username' is required"
+ *         '401':
+ *            description: <b>Unauthorized</b>, with error message.
+ *            content:
+ *              application/json:
+ *               schema:
+ *                 $ref: '#/components/schemas/ErrorMessage'
+ *               examples:
+ *                no token provided:
+ *                 value:
+ *                  message: No authorization token was found.
+ *                user not found:
+ *                 value:
+ *                  message: User not found.
+ *         '403':
+ *          description: <b>Forbidden</b>, with error message.
+ *          content:
+ *           application/json:
+ *            schema:
+ *             $ref: '#/components/schemas/ErrorMessage'
+ *            example:
+ *             message: Not authorized to access this info.
  *         '500':
  *           description: <b>Internal Server Error</b>, with error message.
  *           content:
@@ -112,61 +155,46 @@ const attendanceByUsername = async (req, res) => {
  */
 
 const clockIn = async (req, res) => {
-  try {
-    const { error, value } = joiClockInSchema.validate({ username: req.params.username });
-    if (error) {
-      return res.status(400).json({
-        message: "Validation error.",
-        details: error.details.map((detail) => detail.message),
+  getEmployee(req, res, async (req, res, emp) => {
+    try {
+      const a = await Attendance.findOne({
+        userName: emp.userName,
+        clock_out_time: null,
+        status: "present",
       });
-    }
-    const { username } = value;
 
-    const a = await Attendance.findOne({
-      userName: req.params.username,
-      clock_out_time: null,
-      status: "present",
-    });
+      if (a) {
+        res.status(400).json({
+          message: "Employee must chek-out first.",
+        });
+        return;
+      }
 
-    if (a) {
-      res.status(400).json({
-        message: "Employee must chek-out first.",
+      const attendance = await Attendance.create({
+        userName: emp.userName,
+        clock_in_time: new Date(),
+        clock_out_time: null,
+        status: "present",
       });
-      return;
+
+      res.status(201).json(attendance);
+    } catch (err) {
+      res.status(500).json({ message: err.message });
     }
-
-    const attendance = await Attendance.create({
-      userName: username,
-      clock_in_time: new Date(),
-      clock_out_time: null,
-      status: "present",
-    });
-
-    res.status(201).json(attendance);
-  } catch (err) {
-    res.status(500).json({ message: err.message });
-  }
+  });
 };
 
 /**
  * @openapi
  * paths:
- *   /clockOut/{username}:
+ *   /clockOut:
  *     post:
  *       summary: Clockout employee.
  *       description: Finish attendace with current date-time.
  *       tags:
  *         - Attendance
- *       parameters:
- *         - name: username
- *           in: path
- *           required: true
- *           description: <b>username</b> of the employee
- *           schema:
- *             type: string
- *             minLength: 0
- *             maxLength: 100
- *           example: admin
+ *       security:
+ *        - jwt: []
  *       responses:
  *         '200':
  *           description: <b>OK</b>, with updated attendance.
@@ -182,6 +210,27 @@ const clockIn = async (req, res) => {
  *                 $ref: '#/components/schemas/ErrorMessage'
  *               example:
  *                 message: "Query parameter 'username' is required"
+ *         '401':
+ *            description: <b>Unauthorized</b>, with error message.
+ *            content:
+ *              application/json:
+ *               schema:
+ *                 $ref: '#/components/schemas/ErrorMessage'
+ *               examples:
+ *                no token provided:
+ *                 value:
+ *                  message: No authorization token was found.
+ *                user not found:
+ *                 value:
+ *                  message: User not found.
+ *         '403':
+ *          description: <b>Forbidden</b>, with error message.
+ *          content:
+ *           application/json:
+ *            schema:
+ *             $ref: '#/components/schemas/ErrorMessage'
+ *            example:
+ *             message: Not authorized to access this info.
  *         '500':
  *           description: <b>Internal Server Error</b>, with error message.
  *           content:
@@ -193,39 +242,31 @@ const clockIn = async (req, res) => {
  */
 
 const clockOut = async (req, res) => {
-  try {
-    const { error, value } = joiClockInSchema.validate({ username: req.params.username });
-    if (error) {
-      return res.status(400).json({
-        message: "Validation error.",
-        details: error.details.map((detail) => detail.message),
+  getEmployee(req, res, async (req, res, emp) => {
+    try {
+      const attendace = await Attendance.findOne({
+        userName: emp.userName,
+        clock_out_time: null,
+        status: "present",
       });
+
+      if (!attendace) {
+        res.status(400).json({
+          message: "Employee must chek-in first.",
+        });
+        return;
     }
 
-    const { username } = value;
+      attendace.clock_out_time = new Date();
+      attendace.status = "completed";
 
-    const attendace = await Attendance.findOne({
-      userName: username,
-      clock_out_time: null,
-      status: "present",
-    });
+      const updatedAttendance = await attendace.save();
 
-    if (!attendace) {
-      res.status(400).json({
-        message: "Employee must chek-in first.",
-      });
-      return;
+      res.status(200).json(updatedAttendance);
+    } catch (err) {
+      res.status(500).json({ message: err.message });
     }
-
-    attendace.clock_out_time = new Date();
-    attendace.status = "completed";
-
-    const updatedAttendance = await attendace.save();
-
-    res.status(200).json(updatedAttendance);
-  } catch (err) {
-    res.status(500).json({ message: err.message });
-  }
+  });
 };
 
 /**
@@ -237,6 +278,8 @@ const clockOut = async (req, res) => {
  *       description: Delete **attendance**.
  *       tags:
  *         - Attendance
+ *       security:
+ *        - jwt: []
  *       parameters:
  *         - name: _id
  *           in: path
@@ -262,6 +305,27 @@ const clockOut = async (req, res) => {
  *                 $ref: '#/components/schemas/ErrorMessage'
  *               example:
  *                 message: "Query parameter '_id' is required"
+ *         '401':
+ *            description: <b>Unauthorized</b>, with error message.
+ *            content:
+ *              application/json:
+ *               schema:
+ *                 $ref: '#/components/schemas/ErrorMessage'
+ *               examples:
+ *                no token provided:
+ *                 value:
+ *                  message: No authorization token was found.
+ *                user not found:
+ *                 value:
+ *                  message: User not found.
+ *         '403':
+ *          description: <b>Forbidden</b>, with error message.
+ *          content:
+ *           application/json:
+ *            schema:
+ *             $ref: '#/components/schemas/ErrorMessage'
+ *            example:
+ *             message: Not authorized to access this info.
  *         '404':
  *           description: <b>Not Found</b>, attendance not found.
  *           content:
@@ -279,33 +343,51 @@ const clockOut = async (req, res) => {
  */
 
 const attendaceDeleteOne = async (req, res) => {
-  try {
-    if (!req.params._id || !mongoose.Types.ObjectId.isValid(req.params._id)) {
-      res.status(400).json({
-        message: "Valid query parameter '_id' is required",
+  getEmployee(req, res, async (req, res, emp) => {
+    try {
+      if (!req.params._id || !mongoose.Types.ObjectId.isValid(req.params._id)) {
+        res.status(400).json({
+          message: "Valid query parameter '_id' is required",
+        });
+        return;
+      }
+      if(emp.role != "admin"){
+        return res.status(403).json({
+          message: "Not authorized to make changes.",
+        });
+      }
+      const attendace = await Attendance.findOne({
+        _id: req.params._id,
       });
-      return;
+
+      if (!attendace) {
+        res.status(400).json({
+          message: "Attendance with given '_id' parameter does not exit.",
+        });
+        return;
+      }
+
+      await attendace.deleteOne();
+
+      res.status(200).json({ message: "Attendance successfully deleted." });
+    } catch (err) {
+      res.status(500).json({ message: err.message });
     }
-
-    const attendace = await Attendance.findOne({
-      _id: req.params._id,
-    });
-
-    if (!attendace) {
-      res.status(400).json({
-        message: "Attendance with given '_id' parameter does not exit.",
-      });
-      return;
-    }
-
-    await attendace.deleteOne();
-
-    res.status(200).json({ message: "Attendance successfully deleted." });
-  } catch (err) {
-    res.status(500).json({ message: err.message });
-  }
+  });
 };
 
+
+const getEmployee = async (req, res, cbResult) => {
+  if (req.auth?.userName) {
+    try {
+      let employee = await Employee.findOne({ userName: req.auth.userName }).exec();
+      if (!employee) res.status(401).json({ message: "Not authenticated." });
+      else cbResult(req, res, employee);
+    } catch (err) {
+      res.status(500).json({ message: err.message });
+    }
+  }
+};
 export default {
   attendanceByUsername,
   clockIn,
